REFACTOR: mimic package move for tests
